Fix MailViewer scrolling keys, fix #4018
Page up/Page down/Home/End buttons were not working

It was another case of the MailViewer outlasting it's ViewModel, so references to it's DOM element in the ViewModel were being lost.

We need to be mindful of cases like this, especially when doing these refactorings.

Now scrolling of the MailViewer is handled entirely within the MailViewer itself (registering shortcuts and doing the scrolling).

for bonus points, you can now use these keys to scroll in the search view, since the shortcuts are no longer being setup by the MailView.
